+++ temba/msgs/models.py
@@ -199,12 +199,9 @@ class Broadcast(models.Model):
     translations = models.JSONField()
     base_language = models.CharField(max_length=3)  # ISO-639-3
 
-    channel = models.ForeignKey(Channel, on_delete=models.PROTECT, null=True)
-    ticket = models.ForeignKey("tickets.Ticket", on_delete=models.PROTECT, null=True, related_name="broadcasts")
     status = models.CharField(max_length=1, choices=STATUS_CHOICES, default=STATUS_QUEUED)
-
     created_by = models.ForeignKey(User, null=True, on_delete=models.PROTECT, related_name="broadcast_creations")
-    created_on = models.DateTimeField(default=timezone.now, db_index=True)  # TODO remove index
+    created_on = models.DateTimeField(default=timezone.now)
     modified_by = models.ForeignKey(User, null=True, on_delete=models.PROTECT, related_name="broadcast_modifications")
     modified_on = models.DateTimeField(default=timezone.now)
 
@@ -226,8 +223,6 @@ class Broadcast(models.Model):
         contacts=None,
         urns: list[str] = None,
         contact_ids: list[int] = None,
-        channel: Channel = None,
-        ticket=None,
         **kwargs,
     ):
         # if base language is not provided
@@ -254,8 +249,6 @@ class Broadcast(models.Model):
 
         broadcast = cls.objects.create(
             org=org,
-            channel=channel,
-            ticket=ticket,
             translations=translations,
             base_language=base_language,
             created_by=user,